Fix everything

+++ src/main/java/edu/kit/kastel/sdq/artemis4j/client/StudentExamDTO.java
@@ -1,4 +1,4 @@
-/* Licensed under EPL-2.0 2024-2025. */
+/* Licensed under EPL-2.0 2024-2026. */
 package edu.kit.kastel.sdq.artemis4j.client;
 
 import java.util.Arrays;
@@ -12,14 +12,14 @@ public record StudentExamDTO(
         @JsonProperty boolean submitted,
         @JsonProperty boolean started,
         @JsonProperty UserDTO user) {
-    public static List<StudentExamDTO> fetchAll(ArtemisClient client, int courseId, long examId)
+    public static List<StudentExamDTO> fetchAll(ArtemisClient client, long courseId, long examId)
             throws ArtemisNetworkException {
         return Arrays.asList(ArtemisRequest.get()
                 .path(List.of("exam", "courses", courseId, "exams", examId, "student-exams"))
                 .executeAndDecode(client, StudentExamDTO[].class));
     }
 
-    public static void toggleToSubmitted(ArtemisClient client, int courseId, long examId, long studentExamId)
+    public static void toggleToSubmitted(ArtemisClient client, long courseId, long examId, long studentExamId)
             throws ArtemisNetworkException {
         ArtemisRequest.put()
                 .path(List.of(
